+++ extensions/doctor/src/setOutput.ts
@@ -1,0 +1,8 @@
+import * as vscode from 'vscode';
+
+const channel = vscode.window.createOutputChannel('Doctor');
+
+export default function (message: string) {
+  channel.appendLine(message);
+  channel.show();
+}