+++ extensions/doctor/web/src/pages/Dashboard/components/ScanCard/components/AliEslintReport/index.tsx
@@ -8,18 +8,18 @@ const AliEslintReport = (props) => {
 
   const Description = (
     <p className={styles.description}>
-      {window.USE_EN ? 'Use' : '使用'}
+      {window.USE_EN ? 'Use ' : '使用 '}
       <a href="https://www.npmjs.com/package/@iceworks/spec" target="_blank">
         @iceworks/spec
       </a>
-      {window.USE_EN ? 'scan your code' : '扫描代码'}
+      {window.USE_EN ? ' scan your code' : ' 扫描代码'}
     </p>
   );
 
   return (
     <EslintMessages
       Description={Description}
-      reportKey={getReportKey('aliEslint')}
+      reportKey={getReportKey('ESLint')}
       reports={data.reports}
       score={data.score}
     />