+++ packages/constant/src/index.ts
@@ -33,3 +33,5 @@ export const ALI_DIP_PRO = Base64.decode('aHR0cHM6Ly9hZHAuZmMuYWxpYmFiYS1pbmMuY2
 export const ALI_EXTENSION_UPLOAD_URL = Base64.decode('aHR0cHM6Ly9tYXJrZXRwbGFjZS5hbnRmaW4taW5jLmNvbS9vcGVuYXBpL2V4dGVuc2lvbi91cGxvYWQ=');
 export const ALI_EXTENSION_STATUS_UPDATE_URL = Base64.decode('aHR0cHM6Ly9tYXJrZXRwbGFjZS5hbnRmaW4taW5jLmNvbS9vcGVuYXBpL2V4dGVuc2lvbi9zdGF0dXM=');
 export const ICEWORKS_ICON_PATH = 'https://img.alicdn.com/imgextra/i4/O1CN01AvqMOu1sYpY1j8xaI_!!6000000005779-2-tps-574-204.png';
+
+export const FUSION_MOBILE_MATERIAL_BASE_HOME_URL = Base64.decode('aHR0cHM6Ly9mdXNpb24uYWxpYmFiYS1pbmMuY29tL21vYmlsZQ==');