+++ packages/material-engine/src/material/index.ts
@@ -1,4 +1,5 @@
 import * as kebabCase from 'lodash.kebabcase';
+import * as lowerCase from 'lodash.lowercase';
 import axios from 'axios';
 import {
   checkIsAliInternal,
@@ -6,6 +7,7 @@ import {
   getDataFromSettingJson,
   CONFIGURATION_KEY_MATERIAL_SOURCES,
 } from '@appworks/common-service';
+import { FUSION_MOBILE_MATERIAL_BASE_HOME_URL } from '@appworks/constant';
 import { IMaterialSource, IMaterialData, IMaterialBase } from '@appworks/material-utils';
 import { getProjectType } from '@appworks/project-service';
 import i18n from './i18n';
@@ -14,16 +16,42 @@ import generateDebugMaterialData from './generateDebugMaterialData';
 export const DEBUG_PREFIX = 'DEBUG:';
 export { generateDebugMaterialData };
 
+// material source
 const ICE_MATERIAL_SOURCE = 'https://ice.alicdn.com/assets/materials/react-materials.json';
 const ANTD_MATERIAL_SOURCE = 'https://ice.alicdn.com/assets/materials/antd-materials.json';
 const VUE_MATERIAL_SOURCE = 'https://ice.alicdn.com/assets/materials/vue-materials.json';
 // const MINI_PROGRAM_MATERIAL_SOURCE = 'https://ice.alicdn.com/assets/materials/miniprogram-materials.json';
 const RAX_MATERIAL_SOURCE = 'https://ice.alicdn.com/assets/materials/rax-materials.json';
+// base component source
+const FUSION_PC_COMPONENTS_SOURCE = 'https://ice.alicdn.com/assets/base-components-1.x.json';
+const RAX_BASE_COMPONENTS_SOURCE = 'https://ice.alicdn.com/assets/rax-base-components.json';
+const ANTD_PC_COMPONENTS_SOURCE = 'https://iceworks.oss-cn-hangzhou.aliyuncs.com/assets/antd-components.json';
+// base home url
+const FUSION_PC_MATERIAL_BASE_HOME_URL = 'https://fusion.design/pc/component';
+const ANTD_PC_MATERIAL_BASE_HOME_URL = 'https://ant.design/components';
+// base repository url
+const FUSION_PC_MATERIAL_BASE_REPOSITORY_URL = 'https://github.com/alibaba-fusion/next/tree/master/src';
+const ANTD_PC_MATERIAL_BASE_REPOSITORY_URL = 'https://github.com/ant-design/ant-design/tree/master/components';
 
-const MATERIAL_BASE_HOME_URL = 'https://ice.work/component';
-const MATERIAL_BASE_REPOSITORY_URL = 'https://github.com/alibaba-fusion/next/tree/master/src';
-const ICE_BASE_COMPONENTS_SOURCE = 'https://ice.alicdn.com/assets/base-components-1.x.json';
-const RAX_BASE_COMPONENTS_SOURCE = 'http://ice.alicdn.com/assets/rax-base-components.json';
+const componentSourceDetails = [
+  {
+    npm: '@alifd/next',
+    homeUrl: FUSION_PC_MATERIAL_BASE_HOME_URL,
+    componentNameFormatFunc: kebabCase,
+    repositoryUrl: FUSION_PC_MATERIAL_BASE_REPOSITORY_URL,
+  },
+  {
+    npm: '@alifd/meet',
+    homeUrl: FUSION_MOBILE_MATERIAL_BASE_HOME_URL,
+    componentNameFormatFunc: kebabCase,
+  },
+  {
+    npm: 'antd',
+    homeUrl: ANTD_PC_MATERIAL_BASE_HOME_URL,
+    componentNameFormatFunc: lowerCase,
+    repositoryUrl: ANTD_PC_MATERIAL_BASE_REPOSITORY_URL,
+  },
+];
 
 const OFFICAL_MATERIAL_SOURCES = [
   {
@@ -68,6 +96,10 @@ let dataCache: { [source: string]: IMaterialData } = {};
 
 const isIceMaterial = (source: string) => {
   return source === ICE_MATERIAL_SOURCE;
+};
+
+const isAntdMaterial = (source: string) => {
+  return source === ANTD_MATERIAL_SOURCE;
 };
 
 const isRaxMaterial = (source: string) => {
@@ -137,8 +169,11 @@ export const getData = async function (source: string): Promise<IMaterialData> {
     let bases: IMaterialBase[];
     try {
       if (isIceMaterial(source)) {
-        const baseResult = await axios({ url: ICE_BASE_COMPONENTS_SOURCE });
+        const baseResult = await axios({ url: FUSION_PC_COMPONENTS_SOURCE });
         bases = getBaseMaterials(baseResult.data, '@alifd/next', '1.18.16');
+      } else if (isAntdMaterial(source)) {
+        const baseResult = await axios({ url: ANTD_PC_COMPONENTS_SOURCE });
+        bases = getBaseMaterials(baseResult.data, 'antd', '4.16.5');
       } else if (isRaxMaterial(source)) {
         const baseResult = await axios({ url: RAX_BASE_COMPONENTS_SOURCE });
         bases = baseResult.data;
@@ -162,6 +197,9 @@ export const getData = async function (source: string): Promise<IMaterialData> {
 };
 
 function getBaseMaterials(data, npm?, version?) {
+  const componentSourceDetail = componentSourceDetails.find((item) => item.npm === npm) || {} as any;
+  const { homeUrl, repositoryUrl, componentNameFormatFunc } = componentSourceDetail;
+
   return data.map((base: any) => {
     const { name, title, type, importStatement, source } = base;
     return {
@@ -169,13 +207,13 @@ function getBaseMaterials(data, npm?, version?) {
       title,
       categories: [type],
       importStatement,
-      homepage: `${MATERIAL_BASE_HOME_URL}/${name.toLowerCase()}`,
-      repository: `${MATERIAL_BASE_REPOSITORY_URL}/${kebabCase(name)}`,
+      homepage: homeUrl ? `${homeUrl}/${componentNameFormatFunc(name)}` : '',
+      repository: repositoryUrl ? `${repositoryUrl}/${componentNameFormatFunc(name)}` : '',
       source: source || {
         type: 'npm',
         npm,
         version,
-        registry: 'http://registry.npmjs.com',
+        registry: 'https://registry.npmjs.com',
       },
     };
   });